python_scripts/requirements.txt
beautifulsoup4
bs4          
certifi      
cffi         
charset-normalizer
cryptography     
html2text       
idna            
packaging       
pdf2image       
pdfminer.six    
pdfplumber     
pillow         
pycparser 
PyMuPDF    
PyMuPDFb          
pypandoc          
pypdfium2         
pytesseract       
requests          
soupsieve         
urllib3           
pyyaml